Add files via upload
I finished up the power ups except the increase number of bomb that one im not so sure. We need to find out a way to place the grass by cloning or duplicating which ever you are comfortable with. Idk if it will work or not. (its just a suggestion)